Change Signature into a struct and make partial so different sub classes can be split into their own files.
