+++ lcms2.net/types/Signature.Blank.cs
@@ -1,0 +1,5 @@
+﻿namespace lcms2.types;
+
+public partial struct Signature
+{
+}
